backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
djoser==2.2.0
pillow==10.0.1
django-filter==23.3
drf-extra-fields==3.7.0
python-dotenv==1.0.0
